backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
postgrest
realtime
storage3
gotrue
httpx
APScheduler
pytz
pydantic
pydantic-settings
python-dateutil
psycopg2-binary
requests
cors
